+++ ttp_templates/utils/models.py
@@ -1,4 +1,12 @@
-from pydantic import BaseModel, ConfigDict, StrictInt, StrictStr, StrictBool
+from pydantic import (
+    BaseModel,
+    ConfigDict,
+    StrictBool,
+    StrictFloat,
+    StrictInt,
+    StrictStr,
+    model_validator,
+)
 from typing import Union, List
 
 
@@ -43,9 +51,30 @@ class InterfaceStatusRecord(BaseModel):
     packets_out: Union[None, StrictInt]
     rate_bps_in: Union[None, StrictInt]
     rate_bps_out: Union[None, StrictInt]
+    input_utilization: Union[None, StrictFloat] = None
+    output_utilization: Union[None, StrictFloat] = None
     rate_pps_in: Union[None, StrictInt]
     rate_pps_out: Union[None, StrictInt]
     rate_interval: Union[None, StrictInt]
+
+    @staticmethod
+    def _calculate_utilization(
+        rate_bps: Union[None, StrictInt], speed_bps: Union[None, StrictInt]
+    ) -> Union[None, float]:
+        if rate_bps is None or speed_bps is None or speed_bps <= 0:
+            return None
+        return rate_bps * 100 / speed_bps
+
+    @model_validator(mode="after")
+    def calculate_utilization(self):
+        """Calculate input/output utilization as a percentage of port speed."""
+        self.input_utilization = self._calculate_utilization(
+            self.rate_bps_in, self.speed_bps
+        )
+        self.output_utilization = self._calculate_utilization(
+            self.rate_bps_out, self.speed_bps
+        )
+        return self
 
 
 class LldpNeighborRecord(BaseModel):
@@ -114,6 +143,12 @@ class BgpNeighborRecord(BaseModel):
     ipv6_sr_te_prefixes_received: Union[None, StrictInt] = None
     link_state_prefixes_sent: Union[None, StrictInt] = None
     link_state_prefixes_received: Union[None, StrictInt] = None
+
+
+class BgpCommunityRecord(BaseModel):
+    value: StrictStr
+    type: StrictStr
+    name: StrictStr
 
 
 class InventoryRecord(BaseModel):